+++ lib/src/restrr_base.dart
@@ -42,9 +42,11 @@ class RestrrBuilder {
     Restrr.log.info('Attempting to initialize a session (${initType.name}) with $uri');
     // check if the URI is valid
     final RestResponse<HealthResponse> statusResponse = await Restrr.checkUri(uri);
-    if (!statusResponse.hasData) {
+    if (statusResponse.hasError) {
       Restrr.log.warning('Invalid financrr URI: $uri');
-      return RestrrError.invalidUri.toRestResponse();
+      return statusResponse.error == RestrrError.unknown
+          ? RestrrError.invalidUri.toRestResponse()
+          : statusResponse.error?.toRestResponse() ?? RestrrError.invalidUri.toRestResponse();
     }
     Restrr.log.info('Host: $uri, API v${statusResponse.data!.apiVersion}');
     // create the API instance